Fix(HandCards): Ace 계산 로직 수정
