+++ src/main/java/domain/participant/HandCards.java
@@ -1,0 +1,73 @@
+package domain.participant;
+
+import domain.card.Card;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class HandCards {
+    private List<Card> cards;
+    private static final int BUST_CONDITION = 22;
+
+    public HandCards() {
+        this.cards = new ArrayList<>();
+    }
+
+    public void drawCard(Card card) {
+        cards.add(card);
+    }
+
+    public List<String> cardsToString() {
+        List<String> cardList = new ArrayList<>();
+
+        for (Card card : cards) {
+            cardList.add(card.cardToKorName());
+        }
+
+        return cardList;
+    }
+
+    public int calculateCardsScore(){
+        int aceCount = countAce();
+        int score = calculateCardsScoreExceptAce() + aceCount;
+
+        while(aceCount > 0) {
+            score += determineAce(score);
+            aceCount--;
+        }
+
+        return score;
+    }
+
+    private int calculateCardsScoreExceptAce() {
+        int score = 0;
+
+        for(Card card : cards) {
+            if (!card.isAce()) {
+                score += card.getValue();
+            }
+        }
+
+        return score;
+    }
+
+    private int countAce() {
+        int aceCount = 0;
+
+        for(Card card : cards) {
+            if (card.isAce()) {
+                aceCount++;
+            }
+        }
+
+        return aceCount;
+    }
+
+    private int determineAce(int currentSum){
+        if (currentSum + 10 < BUST_CONDITION) {
+            return 10;
+        }
+
+        return 0;
+    }
+}
