+++ src/test/java/domain/participant/player/HandCardsTest.java
@@ -1,0 +1,4 @@
+import static org.junit.jupiter.api.Assertions.*;
+class HandCardsTest {
+  
+}
